grow-295 removed v-html, moved files around to be more NUXTlike
