+++ src/components/Settings/TwoStepFaq.vue
